tests: replace fake content table seeder
